The PutChar() function overflowed an 8 bit variable. It had a calculation using an 8 bit "page" variable that would overflow if the glyph was larger than 32x32 pixels since 32x32 pixels == 8 pages tall and 32 pages wide 8X32 = 256 which == 0 in an 8 bit variable. The easy solution was to make the variable 16 bits. The code size increased by 20 bytes with no obvious slowdown.

+++ glcd/gText.cpp
@@ -803,7 +803,7 @@ int gText::PutChar(uint8_t c)
 
 	for(uint8_t i=0; i<bytes; i++)	/* each vertical byte */
 	{
-		uint8_t page = i*width;
+		uint16_t page = i*width; // page must be 16 bit to prevent overflow
 		for(uint8_t j=0; j<width; j++) /* each column */
 		{
 			uint8_t data = FontRead(this->Font+index+page+j);
@@ -881,7 +881,7 @@ int gText::PutChar(uint8_t c)
 
 		glcd_Device::GotoXY(this->x, (dy & ~7));
 
-		uint8_t page = p/8 * width;
+		uint16_t page = p/8 * width; // page must be 16 bit to prevent overflow
 
 		for(uint8_t j=0; j<width; j++) /* each column of font data */
 		{
